+++ tests/Tracing/HttpClient/TraceableHttpClientTest.php
@@ -7,6 +7,8 @@ namespace Sentry\SentryBundle\Tests\Tracing\HttpClient;
 use PHPUnit\Framework\Constraint\Callback;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\NullLogger;
 use Sentry\SentryBundle\Tracing\HttpClient\AbstractTraceableResponse;
 use Sentry\SentryBundle\Tracing\HttpClient\TraceableHttpClient;
 use Sentry\State\HubInterface;
@@ -14,6 +16,7 @@ use Sentry\Tracing\Transaction;
 use Sentry\Tracing\TransactionContext;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
+use Symfony\Contracts\Service\ResetInterface;
 
 final class TraceableHttpClientTest extends TestCase
 {
@@ -23,7 +26,7 @@ final class TraceableHttpClientTest extends TestCase
     private $hub;
 
     /**
-     * @var MockObject&HttpClientInterface
+     * @var MockObject&HttpClientInterface&LoggerAwareInterface&ResetInterface
      */
     private $client;
 
@@ -42,7 +45,7 @@ final class TraceableHttpClientTest extends TestCase
     protected function setUp(): void
     {
         $this->hub = $this->createMock(HubInterface::class);
-        $this->client = $this->createMock(HttpClientInterface::class);
+        $this->client = $this->createMock(TestableHttpClientInterface::class);
         $this->httpClient = new TraceableHttpClient($this->client, $this->hub);
     }
 
@@ -91,8 +94,30 @@ final class TraceableHttpClientTest extends TestCase
         ], $spans[1]->getTags());
     }
 
+    public function testSetLoggerShouldBeForwardedToDecoratedInstance(): void
+    {
+        $logger = new NullLogger();
+        $this->client->expects($this->once())
+            ->method('setLogger')
+            ->with($logger);
+
+        $this->httpClient->setLogger($logger);
+    }
+
+    public function testResetCallShouldBeForwardedToDecoratedInstance(): void
+    {
+        $this->client->expects($this->once())
+            ->method('reset');
+
+        $this->httpClient->reset();
+    }
+
     private static function isHttpClientPackageInstalled(): bool
     {
         return interface_exists(HttpClientInterface::class);
     }
+}
+
+interface TestableHttpClientInterface extends HttpClientInterface, LoggerAwareInterface, ResetInterface
+{
 }